some fixes
routes prefix: from tablefills to Pharmacy

cause we cannot use UUID in sqlite is ok to use string, cause we dont want to always insert an uuid, i made it optional by setting a default null, and if it is null it generates an uuid4 and place it in the query.

added an item to response so i might be able to find the pharmacy after its creation.